+++ minecraft-clone/platform.h
@@ -8,7 +8,7 @@
 unsigned int loadTexture(std::string path, bool is_png);
 unsigned int loadCubemap(std::vector<std::string> path);
 
-void initialize_platform(const char* window_title, float _window_width, float _window_height);
+void initialize_platform(const char* window_title, int _window_width, int _window_height);
 void shutdown_platform();
 
 void platform_begin_frame();
